Move roadmap types to its types file
