+++ site/src/app/roadmap/toolbar/ImportTranscriptPopup.tsx
@@ -1,10 +1,10 @@
 import { FC, useState } from 'react';
 import './ImportTranscriptPopup.scss';
 import { Button as Button2, Form, Modal } from 'react-bootstrap';
-import { addRoadmapPlan, RoadmapPlan, selectAllPlans, setPlanIndex } from '../../../store/slices/roadmapSlice';
+import { addRoadmapPlan, selectAllPlans, setPlanIndex } from '../../../store/slices/roadmapSlice';
 import { useAppDispatch, useAppSelector } from '../../../store/hooks';
 import { parse as parseHTML, HTMLElement } from 'node-html-parser';
-import { BatchCourseData, PlannerQuarterData, PlannerYearData } from '../../../types/types';
+import { BatchCourseData, PlannerQuarterData, PlannerYearData, RoadmapPlan } from '../../../types/types';
 import { quarters } from '@peterportal/types';
 import { searchAPIResults } from '../../../helpers/util';
 import { markTransfersAsUnread } from '../../../helpers/transferCredits';
